Feature: Marshal option to disable final control character
The first version of this library was designed to build valid CNAB
files for Bradesco bank [1], and one of Bradesco's requirements was to contain
the final control hexadecimal character 1A. But as this library started being
used by other integrations, the final control character became an issue, as
it doesn't comply with other specifications.

Now it will be possible to disable the final control character. But it will
still be true to keep backward compatibility with the current library
behaviour.

[1] https://banco.bradesco/assets/pessoajuridica/pdf/4008-524-0121-layout-cobranca-versao-portugues.pdf